pyproject.toml
[project]
name = "airflow3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow-providers-fab>=2.4.0",
    "apache-airflow-providers-ftp>=3.13.0",
    "apache-airflow-providers-http>=5.3.0",
    "apache-airflow-providers-mysql>=6.3.5",
    "apache-airflow-providers-postgres>=6.2.0",
    "astronomer-cosmos>=1.10.1",
    "dbt-core>=1.9.6",
    "dbt-postgres>=1.9.0",
    "fhir-resources>=8.0.0",
    "great-expectations>=0.18.22",
    "joblib>=1.5.2",
    "openpyxl>=3.1.5",
    "pandas==2.1.4",
    "pandera>=0.24.0",
    "psycopg2-binary>=2.9.10",
    "pyarrow>=20.0.0",
    "pydbml>=1.2.0",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.5",
    "sqlalchemy>=1.4.54",
    "ujson>=5.10.0",
    "webdav4>=0.10.0",
    "xlsxwriter>=3.2.3",
]

[tool.uv.workspace]
members = [
    "dags/hisp-project-analytics",
    "airflow-env",
]
